# Use float instead of double for constants to improve speed