--- a/src/StepperMotor.cpp
+++ b/src/StepperMotor.cpp
@@ -11,7 +11,7 @@ StepperMotor::StepperMotor(int pp, float _R)
   // save phase resistance number
   phase_resistance = _R;
 
-  // torque control type is voltage by default 
+  // torque control type is voltage by default
   // current and foc_current not supported yet
   torque_controller = TorqueControlType::voltage;
 }
@@ -23,7 +23,7 @@ void StepperMotor::linkDriver(StepperDriver* _driver) {
   driver = _driver;
 }
 
-// init hardware pins   
+// init hardware pins
 void StepperMotor::init() {
   if(monitor_port) monitor_port->println(F("MOT: Init"));
 
@@ -37,7 +37,7 @@ void StepperMotor::init() {
   if(voltage_limit > driver->voltage_limit) voltage_limit =  driver->voltage_limit;
   // constrain voltage for sensor alignment
   if(voltage_sensor_align > voltage_limit) voltage_sensor_align = voltage_limit;
-  
+
   // update the controller limits
   if(_isset(phase_resistance)){
     // velocity control loop controls current
@@ -52,7 +52,7 @@ void StepperMotor::init() {
   if(monitor_port) monitor_port->println(F("MOT: Enable driver."));
   enable();
   _delay(500);
-  
+
 }
 
 
@@ -101,14 +101,14 @@ int  StepperMotor::initFOC( float zero_electric_offset, Direction _sensor_direct
     // added the shaft_angle update
     shaft_angle = sensor->getAngle();
   }else if(monitor_port) monitor_port->println(F("MOT: No sensor."));
-  
+
   if(exit_flag){
     if(monitor_port) monitor_port->println(F("MOT: Ready."));
   }else{
     if(monitor_port) monitor_port->println(F("MOT: Init FOC failed."));
     disable();
   }
- 
+
   return exit_flag;
 }
 
@@ -116,7 +116,7 @@ int  StepperMotor::initFOC( float zero_electric_offset, Direction _sensor_direct
 int StepperMotor::alignSensor() {
   int exit_flag = 1; //success
   if(monitor_port) monitor_port->println(F("MOT: Align sensor."));
-  
+
   // if unknown natural direction
   if(!_isset(sensor_direction)){
     // check if sensor needs zero search
@@ -127,7 +127,7 @@ int StepperMotor::alignSensor() {
     // find natural direction
     // move one electrical revolution forward
     for (int i = 0; i <=500; i++ ) {
-      float angle = _3PI_2 + _2PI * i / 500.0;
+      float angle = _3PI_2 + _2PI * i / 500.0f;
       setPhaseVoltage(voltage_sensor_align, 0,  angle);
       _delay(2);
     }
@@ -135,14 +135,14 @@ int StepperMotor::alignSensor() {
     float mid_angle = sensor->getAngle();
     // move one electrical revolution backwards
     for (int i = 500; i >=0; i-- ) {
-      float angle = _3PI_2 + _2PI * i / 500.0 ;
+      float angle = _3PI_2 + _2PI * i / 500.0f ;
       setPhaseVoltage(voltage_sensor_align, 0,  angle);
       _delay(2);
     }
     float end_angle = sensor->getAngle();
     setPhaseVoltage(0, 0, 0);
     _delay(200);
-    // determine the direction the sensor moved 
+    // determine the direction the sensor moved
     if (mid_angle == end_angle) {
       if(monitor_port) monitor_port->println(F("MOT: Failed to notice movement"));
       return 0; // failed calibration
@@ -153,10 +153,10 @@ int StepperMotor::alignSensor() {
       if(monitor_port) monitor_port->println(F("MOT: sensor_direction==CW"));
       sensor_direction = Direction::CW;
     }
-    // check pole pair number 
+    // check pole pair number
     if(monitor_port) monitor_port->print(F("MOT: PP check: "));
     float moved =  fabs(mid_angle - end_angle);
-    if( fabs(moved*pole_pairs - _2PI) > 0.5 ) { // 0.5 is arbitrary number it can be lower or higher!
+    if( fabs(moved*pole_pairs - _2PI) > 0.5f ) { // 0.5f is arbitrary number it can be lower or higher!
       if(monitor_port) monitor_port->print(F("fail - estimated pp:"));
       if(monitor_port) monitor_port->println(_2PI/moved,4);
     }else if(monitor_port) monitor_port->println(F("OK!"));
@@ -166,7 +166,7 @@ int StepperMotor::alignSensor() {
   // zero electric angle not known
   if(!_isset(zero_electric_angle)){
     // align the electrical phases of the motor and sensor
-    // set angle -90(270 = 3PI/2) degrees 
+    // set angle -90(270 = 3PI/2) degrees
     setPhaseVoltage(voltage_sensor_align, 0,  _3PI_2);
     _delay(700);
     zero_electric_angle = _normalizeAngle(_electricalAngle(sensor_direction*sensor->getAngle(), pole_pairs));
@@ -185,7 +185,7 @@ int StepperMotor::alignSensor() {
 // Encoder alignment the absolute zero angle
 // - to the index
 int StepperMotor::absoluteZeroSearch() {
-  
+
   if(monitor_port) monitor_port->println(F("MOT: Index search..."));
   // search the absolute zero with small velocity
   float limit_vel = velocity_limit;
@@ -194,7 +194,7 @@ int StepperMotor::absoluteZeroSearch() {
   voltage_limit = voltage_sensor_align;
   shaft_angle = 0;
   while(sensor->needsSearch() && shaft_angle < _2PI){
-    angleOpenloop(1.5*_2PI);
+    angleOpenloop(1.5f*_2PI);
     // call important for some sensors not to loose count
     // not needed for the search
     sensor->getAngle();
@@ -218,15 +218,15 @@ int StepperMotor::absoluteZeroSearch() {
 void StepperMotor::loopFOC() {
   // if open-loop do nothing
   if( controller==MotionControlType::angle_openloop || controller==MotionControlType::velocity_openloop ) return;
-  // shaft angle 
+  // shaft angle
   shaft_angle = shaftAngle();
-  
+
   // if disabled do nothing
-  if(!enabled) return; 
+  if(!enabled) return;
 
   electrical_angle = electricalAngle();
 
-  // set the phase voltage - FOC heart function :) 
+  // set the phase voltage - FOC heart function :)
   setPhaseVoltage(voltage.q, voltage.d, electrical_angle);
 }
 
@@ -239,7 +239,7 @@ void StepperMotor::move(float new_target) {
   // get angular velocity
   shaft_velocity = shaftVelocity();
   // if disabled do nothing
-  if(!enabled) return; 
+  if(!enabled) return;
   // downsampling (optional)
   if(motion_cnt++ < motion_downsample) return;
   motion_cnt = 0;
@@ -249,7 +249,7 @@ void StepperMotor::move(float new_target) {
   switch (controller) {
     case MotionControlType::torque:
       if(!_isset(phase_resistance))  voltage.q = target; // if voltage torque control
-      else voltage.q =  target*phase_resistance; 
+      else voltage.q =  target*phase_resistance;
       voltage.d = 0;
       break;
     case MotionControlType::angle:
@@ -259,7 +259,7 @@ void StepperMotor::move(float new_target) {
       shaft_velocity_sp = P_angle( shaft_angle_sp - shaft_angle );
       // calculate the torque command
       current_sp = PID_velocity(shaft_velocity_sp - shaft_velocity); // if voltage torque control
-      // if torque controlled through voltage  
+      // if torque controlled through voltage
       // use voltage if phase-resistance not provided
       if(!_isset(phase_resistance))  voltage.q = current_sp;
       else  voltage.q = current_sp*phase_resistance;
@@ -270,7 +270,7 @@ void StepperMotor::move(float new_target) {
       shaft_velocity_sp = target;
       // calculate the torque command
       current_sp = PID_velocity(shaft_velocity_sp - shaft_velocity); // if current/foc_current torque control
-      // if torque controlled through voltage control 
+      // if torque controlled through voltage control
       // use voltage if phase-resistance not provided
       if(!_isset(phase_resistance))  voltage.q = current_sp;
       else  voltage.q = current_sp*phase_resistance;
@@ -295,12 +295,12 @@ void StepperMotor::move(float new_target) {
 // Method using FOC to set Uq and Ud to the motor at the optimal angle
 // Function implementing Sine PWM algorithms
 // - space vector not implemented yet
-// 
+//
 // Function using sine approximation
 // regular sin + cos ~300us    (no memory usaage)
 // approx  _sin + _cos ~110us  (400Byte ~ 20% of memory)
 void StepperMotor::setPhaseVoltage(float Uq, float Ud, float angle_el) {
-  // Sinusoidal PWM modulation 
+  // Sinusoidal PWM modulation
   // Inverse Park transformation
 
   // angle normalization in between 0 and 2pi
@@ -323,18 +323,18 @@ float StepperMotor::velocityOpenloop(float target_velocity){
   // get current timestamp
   unsigned long now_us = _micros();
   // calculate the sample time from last call
-  float Ts = (now_us - open_loop_timestamp) * 1e-6;
+  float Ts = (now_us - open_loop_timestamp) * 1e-6f;
   // quick fix for strange cases (micros overflow + timestamp not defined)
-  if(Ts <= 0 || Ts > 0.5) Ts = 1e-3; 
+  if(Ts <= 0 || Ts > 0.5f) Ts = 1e-3f;
 
   // calculate the necessary angle to achieve target velocity
   shaft_angle = _normalizeAngle(shaft_angle + target_velocity*Ts);
   // for display purposes
   shaft_velocity = target_velocity;
-    
+
   // use voltage limit or current limit
   float Uq = voltage_limit;
-  if(_isset(phase_resistance)) Uq =  current_limit*phase_resistance; 
+  if(_isset(phase_resistance)) Uq =  current_limit*phase_resistance;
 
   // set the maximal allowed voltage (voltage_limit) with the necessary angle
   setPhaseVoltage(Uq,  0, _electricalAngle(shaft_angle, pole_pairs));
@@ -352,9 +352,9 @@ float StepperMotor::angleOpenloop(float target_angle){
   // get current timestamp
   unsigned long now_us = _micros();
   // calculate the sample time from last call
-  float Ts = (now_us - open_loop_timestamp) * 1e-6;
+  float Ts = (now_us - open_loop_timestamp) * 1e-6f;
   // quick fix for strange cases (micros overflow + timestamp not defined)
-  if(Ts <= 0 || Ts > 0.5) Ts = 1e-3; 
+  if(Ts <= 0 || Ts > 0.5f) Ts = 1e-3f;
 
   // calculate the necessary angle to move from current position towards target angle
   // with maximal velocity (velocity_limit)
@@ -368,12 +368,12 @@ float StepperMotor::angleOpenloop(float target_angle){
 
   // use voltage limit or current limit
   float Uq = voltage_limit;
-  if(_isset(phase_resistance)) Uq =  current_limit*phase_resistance; 
+  if(_isset(phase_resistance)) Uq =  current_limit*phase_resistance;
   // set the maximal allowed voltage (voltage_limit) with the necessary angle
   setPhaseVoltage(Uq,  0, _electricalAngle(shaft_angle, pole_pairs));
 
   // save timestamp for next call
   open_loop_timestamp = now_us;
-  
+
   return Uq;
 }