--- a/src/common/base_classes/Sensor.cpp
+++ b/src/common/base_classes/Sensor.cpp
@@ -2,7 +2,7 @@
 #include "../foc_utils.h"
 #include "../time_utils.h"
 
- /** 
+ /**
  * returns 0 if it does need search for absolute zero
  * 0 - magnetic sensor (& encoder with index which is found)
  * 1 - ecoder with index (with index not found yet)
@@ -16,15 +16,15 @@ float Sensor::getVelocity(){
 
     // calculate sample time
     unsigned long now_us = _micros();
-    float Ts = (now_us - velocity_calc_timestamp)*1e-6;
+    float Ts = (now_us - velocity_calc_timestamp)*1e-6f;
     // quick fix for strange cases (micros overflow)
-    if(Ts <= 0 || Ts > 0.5) Ts = 1e-3; 
+    if(Ts <= 0 || Ts > 0.5f) Ts = 1e-3f;
 
     // current angle
     float angle_c = getAngle();
     // velocity calculation
     float vel = (angle_c - angle_prev)/Ts;
-    
+
     // save variables for future pass
     angle_prev = angle_c;
     velocity_calc_timestamp = now_us;