--- a/src/common/pid.cpp
+++ b/src/common/pid.cpp
@@ -6,9 +6,9 @@ PIDController::PIDController(float P, float I, float D, float ramp, float limit)
     , D(D)
     , output_ramp(ramp)    // output derivative limit [volts/second]
     , limit(limit)         // output supply limit     [volts]
-    , integral_prev(0.0)
-    , error_prev(0.0)
-    , output_prev(0.0)
+    , integral_prev(0.0f)
+    , error_prev(0.0f)
+    , output_prev(0.0f)
 {
     timestamp_prev = _micros();
 }
@@ -17,13 +17,13 @@ PIDController::PIDController(float P, float I, float D, float ramp, float limit)
 float PIDController::operator() (float error){
     // calculate the time from the last call
     unsigned long timestamp_now = _micros();
-    float Ts = (timestamp_now - timestamp_prev) * 1e-6;
+    float Ts = (timestamp_now - timestamp_prev) * 1e-6f;
     // quick fix for strange cases (micros overflow)
-    if(Ts <= 0 || Ts > 0.5) Ts = 1e-3; 
+    if(Ts <= 0 || Ts > 0.5f) Ts = 1e-3f;
 
     // u(s) = (P + I/s + Ds)e(s)
     // Discrete implementations
-    // proportional part 
+    // proportional part
     // u_p  = P *e(k)
     float proportional = P * error;
     // Tustin transform of the integral part
@@ -56,4 +56,3 @@ float PIDController::operator() (float error){
     timestamp_prev = timestamp_now;
     return output;
 }
-