--- a/src/current_sense/hardware_api.h
+++ b/src/current_sense/hardware_api.h
@@ -6,14 +6,14 @@
 
 /**
  *  function reading an ADC value and returning the read voltage
- * 
+ *
  * @param pinA - the arduino pin to be read (it has to be ADC pin)
  */
 float _readADCVoltageInline(const int pinA);
 
 /**
  *  function reading an ADC value and returning the read voltage
- * 
+ *
  * @param pinA - adc pin A
  * @param pinB - adc pin B
  * @param pinC - adc pin C
@@ -21,8 +21,8 @@ float _readADCVoltageInline(const int pinA);
 void _configureADCInline(const int pinA,const int pinB,const int pinC = NOT_SET);
 
 /**
- *  function reading an ADC value and returning the read voltage 
- * 
+ *  function reading an ADC value and returning the read voltage
+ *
  * @param pinA - adc pin A
  * @param pinB - adc pin B
  * @param pinC - adc pin C
@@ -33,7 +33,7 @@ void _startADC3PinConversionLowSide();
 
 /**
  *  function reading an ADC value and returning the read voltage
- * 
+ *
  * @param pinA - the arduino pin to be read (it has to be ADC pin)
  */
 float _readADCVoltageLowSide(const int pinA);