… examples/hardware_specific_examples/Bluepill_examples/encoder/bluepill_position_control/bluepill_position_control.ino
… examples/hardware_specific_examples/Bluepill_examples/magnetic_sensor/bluepill_position_control/bluepill_position_control.ino
… examples/hardware_specific_examples/DRV8302_driver/3pwm_example/encoder/full_control_serial/full_control_serial.ino
… examples/hardware_specific_examples/DRV8302_driver/6pwm_example/encoder/full_control_serial/full_control_serial.ino
… examples/hardware_specific_examples/ESP32/encoder/esp32_position_control/esp32_position_control.ino
… examples/hardware_specific_examples/ESP32/magnetic_sensor/esp32_position_control/esp32_position_control.ino
… examples/hardware_specific_examples/HMBGC_example/position_control/position_control.ino
… examples/hardware_specific_examples/SAMD_examples/nano33IoT/nano33IoT_velocity_control/nano33IoT_velocity_control.ino
… examples/hardware_specific_examples/SimpleFOC-PowerShield/version_v02/single_full_control_example/single_full_control_example.ino
… examples/hardware_specific_examples/SimpleFOCShield/version_v1/double_full_control_example/double_full_control_example.ino
… examples/hardware_specific_examples/SimpleFOCShield/version_v1/single_full_control_example/single_full_control_example.ino
… examples/hardware_specific_examples/SimpleFOCShield/version_v2/double_full_control_example/double_full_control_example.ino
… examples/hardware_specific_examples/SimpleFOCShield/version_v2/single_full_control_example/single_full_control_example.ino
… examples/motion_control/position_motion_control/encoder/angle_control/angle_control.ino
… examples/motion_control/position_motion_control/hall_sensor/angle_control/angle_control.ino
… examples/motion_control/position_motion_control/magnetic_sensor/angle_control/angle_control.ino
… examples/motion_control/torque_control/encoder/current_control/current_control.ino
… examples/motion_control/velocity_motion_control/encoder/velocity_control/velocity_control.ino
… examples/motion_control/velocity_motion_control/hall_sensor/velocity_control.ino
… examples/motion_control/velocity_motion_control/magnetic_sensor/velocity_control/velocity_control.ino
… examples/motor_commands_serial_examples/encoder/full_control_serial/full_control_serial.ino
… examples/motor_commands_serial_examples/hall_sensor/full_control_serial/full_control_serial.ino
… examples/motor_commands_serial_examples/magnetic_sensor/full_control_serial/full_control_serial.ino
… examples/osc_control_examples/osc_esp32_3pwm/osc_esp32_3pwm.ino
… examples/osc_control_examples/osc_esp32_fullcontrol/osc_esp32_fullcontrol.ino
… examples/utils/calibration/alignment_and_cogging_test/alignment_and_cogging_test.ino
… examples/utils/calibration/find_pole_pair_number/encoder/find_pole_pairs_number/find_pole_pairs_number.ino
… examples/utils/calibration/find_pole_pair_number/magnetic_sensor/find_pole_pairs_number/find_pole_pairs_number.ino
… examples/utils/communication_test/step_dir/step_dir_listener_software_interrupt/step_dir_listener_software_interrupt.ino
… examples/utils/communication_test/step_dir/step_dir_motor_example/step_dir_motor_example.ino