We should use float constants instead of double (like `0.0123f`) to improve claculation speed. If we have double constant all expression will be calculated as double just to convert it to float at the end.

I have created a simple benchmark:
```c++
#include <SimpleFOC.h>

#define N 5000UL
#define BENCH(foo, real_foo, n) bench(foo, real_foo, n, #foo)

static float bench_calibration = 0.0f;

float empty(float x)
{
    return 0.0f;
}

float bench_time(float (*foo)(float), size_t n)
{
    uint32_t begin = micros();

    for (size_t i = 0; i < n; i++)
    {
        foo((float)i / n * _2PI);
    }

    uint32_t end = micros();

    return (float)(end - begin) / n - bench_calibration;
}

float bench_error(float (*foo)(float), float (*real_foo)(float), size_t n)
{
    float sum_rms = 0.0f;
    
    for(size_t i = 0; i < n; i++)
    {
        float angle = (float)i / n * _2PI;
        float real = real_foo(angle);
        float calc = foo(angle);
        float diff = calc - real;
        sum_rms += diff * diff;
    }

    return sqrtf(sum_rms / n);
}

void bench(float (*foo)(float), float (*real_foo)(float), size_t n, const char * name)
{

    Serial.print("Testing ");
    Serial.print(name);
    Serial.print(" ");
    Serial.print(n);
    Serial.println(" times");

    float time_spent_us = bench_time(foo, n);
    float rms = bench_error(foo, real_foo, n);

    Serial.print("Time: ");
    Serial.print(time_spent_us, 3);
    Serial.print("us, RMS error: ");
    Serial.println(rms, 6);
}

void bench_calibrate(size_t n)
{
    bench_calibration = 0.0f;
    bench_calibration = bench_time(empty, n);
}

void setup() {
    Serial.begin(115200);

    bench_calibrate(N);

    BENCH(_sin, sinf, N);
    BENCH(_cos, cosf, N);
}

void loop() {

}
```

On STM32F0 running at 48MHz it gave me results presented below.
Before change to floats, using doubles as constants:
```bash
Testing _sin 5000 times
Time: 34.277us, RMS error: 0.001612
Testing _cos 5000 times
Time: 46.258us, RMS error: 0.001612
```

After change to floats:
```bash
Testing _sin 5000 times
Time: 18.794us, RMS error: 0.001612
Testing _cos 5000 times
Time: 24.926us, RMS error: 0.001612
```

As you can see it's almost two times faster and we are talking about 32bit MCU, on 8bit Arduino it should improve even more.

I haven't touched MCU specific files but I can see some double constants here as well.