@runger1101001 2021-05-31
This is a nice idea! And nice test. May I ask, did you also test the changed code with a motor?

I would accept this change, but not against the master branch. Would you be prepared to send a pull request against the dev branch? If not it is not a problem, I can merge your changes over locally to the dev branch, but that would lose the "connection" to your pull request...

Thanks a lot!