@KaSroka 2021-05-31
Hi, yes I will modify the PR to target the dev branch.

I have tested this with stepper motor and it allowed me to achieve better max RPM (600RPM increased to 765RPM)