@langwadt 2021-05-31
-fsingle-precision-constant -Wdouble-promotion

on AVR it shouldn't make any difference because doubles are defined as floats