@KaSroka 2021-05-31
I agree with `-fsingle-precision-constant` but this is something we cannot enforce here, so it's better to sue correct types from the beginning. It will also make huge difference on MCUs with FPU when compiled without `-fsingle-precision-constant`.

I know this touches a lot of files so it may need a lot of testing.