@langwadt 2021-05-31
yeh, if you can't add compiler options that's out, and it would only deal with constants, it won't replace things like sin() with sinf() if any of those functions are used