@runger1101001 2021-06-04
So on SAMD21 I'm getting about 19% improvement on my loopFOC()...

I think I'm in favour of merging this, even though it touches so many files. I was able to apply it cleanly to my local branch.