@askuric 2021-06-05
That's great to hear, well done @KaSroka  for this pull request.

I'm gonna test it today as well and we will merge it probably as is. I just need to make sure this does not break anything for hardware specific implementations. 

I actually think that these changes will impact the least the Arduino, because as far as I know they do not have double at all as a type. 

BTW. we do not use sine functions `sin` or `sinf` in simplefoc. We've got much faster array based implementation `_sin`. 
Also for square root we use the fast inverse square root.