@askuric 2021-06-05
So the results of the testing say:
SimpleFOC PowerShield + Gimbal motor + encoder:
All the times are given in milliseconds and averaged over 1000 `move()` + `loopFOC()` calls.

MCU | `voltage` before| `voltage` after | `foc current` before |`foc current` after  | % gain
--- | ---- | ---| ---- | ---- | ---
ESP32 (r1 d32) | 33-45 us | 26-30us | 79-105 us| 66-82us |  ~30%
STM32 (f411re) | 79-86 us | 50 us | 243-257 us | 177 - 200 us | ~30%
SAMD51 (Metro m4) | 50-55 us | 40-45 us | 143-150 us | 104-143 us | ~20%
Arduino DUE|  174-220us | 160-195 us |  360-430us | 330-400 us | ~10%
Arduino UNO | 1000-1700us | same | 2100-3000us | same | ~0%

So overall the gain is pretty high, so we'll definitely try to make this a standard :D